Development

+++ app/api/art_listings.py
@@ -23,18 +23,18 @@ def get_a_single_listing(artlisting_id):
         return {"errors": "Artlisting not found."}, 404
 
     if request.method == "DELETE":
-        # if not single_listing.check_owner(owner_id):
-        #     return {"errors": "Forbidden."}, 403
-        # else:
-        db.session.delete(single_listing)
-        db.session.commit()
-        # return {"Success": "Listing deleted."}, 202
+        if not single_listing.check_owner(owner_id):
+            return {"errors": "Forbidden."}, 403
+        else:
+            db.session.delete(single_listing)
+            db.session.commit()
+            return {"Success": "Listing deleted."}, 202
 
     if request.method == "PUT":
         form = ArtlistingForm()
         form["csrf_token"].data = request.cookies["csrf_token"]
         if form.validate_on_submit():
-            single_listing.price = form.data["price"]
+            single_listing.price = f"{form.data['price']}"
             single_listing.amount_available = form.data["amount_available"]
             db.session.commit()
             return single_listing.to_safe_dict(), 200
